refactor(oauth2): move generate_code_challenge util to oauth2 utils
